Refactor graphql plugin structure
